Desktop/eros-scheduling/requirements.txt
# EROS Max AI System - Python Dependencies
# Python 3.11+

# Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-monitoring>=2.15.1
google-auth>=2.23.0

# Data Science & ML
pandas>=2.0.3
numpy>=1.24.3
scikit-learn==1.3.0
scipy>=1.11.1

# Utilities
python-dateutil>=2.8.2
pyyaml>=6.0.1

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
